// src/schemas/USSensorFront.ts
// Generated by https://github.com/foxglove/foxglove-sdk
// Options: {}

import { Header } from "@/schemas/Header";

/** Ultrasonic sensor front */
export type USSensorFront = {
  /** Standard ROS header */
  header: Header;

  u_s_sensor4: number;
  u_s_sensor3: number;
  u_s_sensor2: number;
  u_s_sensor1: number;
};