// src/schemas/USSensorRear.ts
// Generated by https://github.com/foxglove/foxglove-sdk
// Options: {}

import { Header } from "@/schemas/Header";

/** Ultrasonic sensor rear */
export type USSensorRear = {
  /** Standard ROS header */
  header: Header;

  u_s_sensor8: number;
  u_s_sensor7: number;
  u_s_sensor6: number;
  u_s_sensor5: number;
};